% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_paf.R
\name{dotplot_paf}
\alias{dotplot_paf}
\title{dotplot_paf genome-wide dotplotter}
\usage{
dotplot_paf(
  paf,
  gapProp = 0.05,
  round2q = 100,
  round2t = round2q,
  maxN2show = NULL,
  pointAlpha = 1,
  gridsEvery = 1e+07,
  minN2plot = 0,
  forceHeatmap = FALSE,
  fixedAxisCoords = TRUE
)
}
\arguments{
\item{paf}{data.table containing the paf-formatted mappings}

\item{gapProp}{numeric [0-1] giving the size of the gaps between chromosomes
as the fraction of the size of the largest chromosome}

\item{round2q}{integer > 0 specifying the rounding factor to use for the
query bins}

\item{round2t}{integer > 0 specifying the rounding factor to use for the
target bins}

\item{maxN2show}{integer, specifying the top of the color scale. Any bins
with more than this number of hits are white.}

\item{pointAlpha}{numeric specifying the opacity of the points}

\item{gridsEvery}{numeric, how many bases should be between gridlines?}

\item{minN2plot}{integer, specifying the minimum number of hits in a bin
for that bin to be plotted.}

\item{forceHeatmap}{logical, should the heatmap be plotted no matter what?}

\item{fixedAxisCoords}{logical, should the coordinates be fixed?}
}
\description{
\code{dotplot_paf} lightweight and fast genome-wide dotplotter.
}
\details{
coming soon
}
