% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_windows.R
\name{minimap2_windows}
\alias{minimap2_windows}
\title{minimap2_windows}
\usage{
minimap2_windows(
  queryFasta,
  targetFasta,
  pafFile,
  windowSize,
  filterTopReps = 1e-04,
  maxTopHits = 50,
  minSecRat = 0.9,
  mismatchPenalty = 4,
  kmerSize = 25,
  kmerStep = 20,
  nCores = 1,
  overwrite = FALSE,
  basicParams = "-A1 -U50,500 --no-long-join -r0,0 --rmq=no -E2,1 -n1 -m0 --frag=yes",
  minimap2call = "minimap2",
  verbose = TRUE
)
}
\arguments{
\item{queryFasta}{character string coercible to a file path specifying the
location of the query (to-be-windowed) fasta file}

\item{targetFasta}{character string coercible to a file path specifying the
location of the target fasta file}

\item{pafFile}{character string coercible to a file path specifying the
location of the output (paf) file}

\item{windowSize}{integer, the basepair size of the window in the query
genome}

\item{filterTopReps}{integer, passed to minimap2 -f. Setting to 0 can
significantly slow operations but will allow visualization of all repeats}

\item{maxTopHits}{integer, passed to minimap2 -N}

\item{minSecRat}{numeric, passed to minimap2 -p}

\item{mismatchPenalty}{numeric, passed to minimap2 -B}

\item{kmerSize}{integer, passed to minimap2 -k}

\item{kmerStep}{integer, passed to minimap2 -w}

\item{nCores}{integer, passed to minimap2 -t}

\item{overwrite}{logical, should existing files be overwritten?}

\item{basicParams}{character string of additional parameters passed to
minimap2.}

\item{verbose}{logical, should updates be printed to the console?}
}
\description{
\code{minimap2_windows} minimap2_windows
}
\details{
coming soon
}
