% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_utils}
\alias{plot_utils}
\alias{order_bySynteny}
\alias{transform_paf2riparian}
\alias{round_rect}
\alias{transform_paf2sameRange}
\alias{scale_between}
\alias{add_alpha}
\alias{are_colors}
\title{Generic internal functions used by DEEPSPACE}
\usage{
order_bySynteny(paf, chrordTarget)

transform_paf2riparian(
  paf,
  chrordQuery,
  chrordTarget,
  gapProp = 0.05,
  howRound = 0.5,
  yQuery = 0,
  yTarget = 1,
  braidOffset = 0.1,
  xoffsetQuery = 0,
  xoffsetTarget = 0,
  sameXrange = FALSE,
  colorColumn = NULL,
  chrWidth = braidOffset
)

round_rect(
  xleft,
  ybottom,
  xright,
  ytop,
  plotWidth,
  plotHeight,
  xrange,
  yrange,
  npts = 50
)

transform_paf2sameRange(paf, xrange)

scale_between(x, min, max, scale1toMean = TRUE)

add_alpha(col, alpha = 1)

are_colors(col)
}
\arguments{
\item{paf}{data.table containing a windowed paf alignment file}

\item{chrordTarget}{vector of chromosome names in the order to be plotted. For
the target genome}

\item{chrordQuery}{vector of chromosome names in the order to be plotted. For
the query genome}

\item{gapProp}{numeric (0-1) giving the proportion of the x axis that should
be gaps between chromosomes}

\item{howRound}{numeric (0-Inf) specifying how rouded the chromosome
polygons should be.}

\item{yQuery}{numeric, y-value of the query genome}

\item{yTarget}{numeric, y-value of the target genome}

\item{braidOffset}{distance between the chromosome midpoint and the y-start
of each daisychain braid. If the same value as chrWidth, the braid starts
exactly at the edge of the chromosome polygon}

\item{xoffsetQuery}{numeric, how far to the left should the query genome be
offset?}

\item{xoffsetTarget}{numeric, how far to the left should the target genome be
offset?}

\item{sameXrange}{logical, should both the query and target genome have the
same x range?}

\item{colorColumn}{character, matching one of the paf columns that contains
the color for that block}

\item{chrWidth}{numeric, providing the thinkness of the chromosomes}

\item{xleft}{left x-position}

\item{ybottom}{bottom y-position}

\item{xright}{right x-position}

\item{ytop}{top y-position}

\item{plotWidth}{width of the plot}

\item{plotHeight}{height of the plot}

\item{xrange}{range of x-values}

\item{yrange}{range of y-values}

\item{npts}{number of points that make up the polygon}

\item{x}{value to transform}

\item{min}{numeric, minimum value}

\item{max}{numeric, maximum value}

\item{scale1toMean}{logical, should the mean be scaled to 1?}

\item{col}{color}

\item{alpha}{opacity
If called, \code{plot_utils} returns its own arguments.}

\item{blks}{a paf file read into R with the following additional fields:
query: name of the query genome, target: name of the target genome, qy: y
coordinate of the the query genome, ty: y coordinate of the target genome,
color: the name of the color for the alignment. This object should contain
all of the pafs for each pairwise contrast together.}
}
\description{
\code{plot_utils} Convenience functions for DEEPSPACE, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{order_bySynteny} order_bySynteny

\code{transform_paf2riparian} transform_paf2riparian

\code{round_rect} from x-y coordinates, make a rounded rectangle

\code{transform_paf2sameRange} transform_paf2sameRange

\code{scale_between} scale_between

\code{add_alpha} add transparency to a color

\code{are_colors} check if a vector is coercible to R colors
}
