% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fastas.R
\name{prep_clwFastas}
\alias{prep_clwFastas}
\alias{prep_fastas}
\title{prep_fastas}
\usage{
prep_fastas(
  queryFasta,
  targetFasta,
  outQueryFasta,
  outTargetFasta,
  windowSize,
  stepSize,
  nWindows,
  stripChrname,
  minChrLen
)
}
\arguments{
\item{queryFasta}{character string coercible to a file path specifying the
location of the query (to-be-windowed) fasta file}

\item{targetFasta}{character string coercible to a file path specifying the
location of the target fasta file}

\item{windowSize}{integer, the basepair size of the window in the query
genome}

\item{stepSize}{integer, the basepair distance between adjacet window start
sites. Defaults to NULL and is calculated from the genome size / nWindows}

\item{nWindows}{integer, the number of windows to chop the query genome.}

\item{stripChrname}{character string coercible to a regular expression passed
to gsub that re-names the sequences in both the query and target fastas}

\item{minChrLen}{integer, minimum length of a sequence to be considered}

\item{queryID}{character string to name the query (windowed) genome}

\item{targetID}{character string to name the target genome}

\item{outDir}{character string coercible to a file path specifying the
output directory}

\item{overwrite}{logical, should existing files be overwritten?}
}
\description{
\code{prep_fastas} prep_fastas
}
\details{
coming soon
}
