% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riparian_paf.R
\name{riparian_paf}
\alias{riparian_paf}
\alias{riparian_blksSameWidth}
\alias{riparian_blksSameScale}
\title{Generic internal functions used by DEEPSPACE}
\usage{
riparian_paf(
  pafFiles,
  labelTheseGenomes = NULL,
  refGenome = NULL,
  genomeIDs,
  sameChrOnly = FALSE,
  round2join = 10000,
  braidColors = c("#C4645C", "#F5915C", "#FFC765", "#FCF8D5", "#BEF3F9", "#66B8FF",
    "#6666FF", "#9C63E1", "#F4BDFF"),
  chrWidth = 0.05,
  sameScale = TRUE,
  orderyBySynteny = TRUE,
  braidAlpha = 0.5,
  braidOffset = 0.06,
  highlightInversions = FALSE,
  braidOutline = NA,
  chrOutlineColor = "grey",
  chrOutlineWidth = 0.1,
  braidOutlineWidth = 0,
  gapProp = 0.05,
  howRound = 0.2
)

riparian_blksSameWidth(
  blks,
  chrList,
  labelTheseGenomes,
  colorColumn = "color",
  braidAlpha = 0.5,
  braidOutline = NA,
  braidOutlineWidth = 0,
  braidPalette = gs_colors,
  chrOutlineColor = "grey",
  chrOutlineWidth = 0.1,
  gapProp = 0.05,
  howRound = 0.2,
  braidOffset = 0.12,
  chrWidth = 0.1
)

riparian_blksSameScale(
  blks,
  chrList,
  labelTheseGenomes,
  chrOutlineColor = "grey",
  chrOutlineWidth = 0.1,
  colorColumn = "color",
  braidAlpha = 0.5,
  braidOutline = NA,
  braidOutlineWidth = 0,
  braidPalette = gs_colors,
  gapProp = 0.05,
  howRound = 0.2,
  braidOffset = 0.12,
  chrWidth = 0.1
)
}
\arguments{
\item{pafFiles}{character vector coercible to a file.path pointing to the
paf files. These must be named as '$QUERYGENOMEID'__s__'$TARGETGENOMEID'.paf.}

\item{labelTheseGenomes}{character string specifying which of the genomeIDs
should have chromosomes labeled in the plot}

\item{refGenome}{one of the genomeIDs that is the reference to which all
alignments are phased. If absent, sameChrOnly should be TRUE}

\item{genomeIDs}{character vector specifying the top-to-bottom genome order.
Must specify a daisy chain where all edges in the vector have a corresponding
.paf.}

\item{sameChrOnly}{logical, should the search only be conducted on alignments
where the query and target sequence names are identical? This speeds things
up, but should only be used in very similar genomes where inter-chromosomal
SVs are known to not exist.}

\item{round2join}{integer specifying the amount of fuzziness should be
present in the joining of alignments against the reference and within the
daisy chain. Larger numbers means less precision, but likely more complete
phasing.}

\item{braidColors}{vector of colors used in the palette for the braid colors.
If a single value, all blocks are not phased against the reference genome.}

\item{chrWidth}{numeric, providing the thinkness of the chromosomes}

\item{sameScale}{logical, do all genomes have the same scale (TRUE) or all
have the same width (FALSE)}

\item{braidAlpha}{numeric (0-1) specifying the opacity of the braids}

\item{braidOffset}{distance between the chromosome midpoint and the y-start
of each daisychain braid. If the same value as chrWidth, the braid starts
exactly at the edge of the chromosome polygon}

\item{highlightInversions}{either NULL or a color specifying what color the
inverted sequence should be}

\item{braidOutline}{color (or NA) specifying the color of the braid outlines}

\item{chrOutlineColor}{color (or NA) specifying the color of the chr outlines}

\item{chrOutlineWidth}{numeric, giving the thickness of the chr outlines}

\item{braidOutlineWidth}{numeric, giving the thickness of the braid outlines}

\item{gapProp}{numeric (0-1) giving the proportion of the x axis that should
be gaps between chromosomes}

\item{howRound}{numeric (0-Inf) specifying how rouded the chromosome
polygons should be.

\cr
If called, \code{riparian_paf} returns its own arguments.}

\item{orderBySynteny}{logical, should chromosomes be ordered by synteny
(TRUE, default) or by the name of chromosomes (FALSE)}
}
\description{
\code{riparian_paf} Convenience functions for DEEPSPACE, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{riparian_paf} riparian_paf

\code{riparian_blksSameWidth} riparian_blksSameWidth

\code{riparian_blksSameScale} riparian_blksSameScale
}
