% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny_paf.R
\name{synteny_paf}
\alias{synteny_paf}
\alias{choose_topNmatch}
\alias{drop_repetitiveTargets}
\alias{choose_localUniqHits}
\alias{choose_topWindowHits}
\alias{project_anchorsOnPaf}
\alias{choose_localBestHits}
\alias{mcscanx_paf}
\alias{calc_pafBlkCoords}
\title{synteny inference from windowed alignments}
\usage{
choose_topNmatch(paf, mapq20nmatchQuantile = 0.1, mapq60nmatchQuantile = 0)

drop_repetitiveTargets(paf, round2, targetNhitsQuantile)

choose_localUniqHits(paf, round2)

choose_topWindowHits(paf, maxBestMaps = 5, topNhits2keep = 2)

project_anchorsOnPaf(anchorPaf, paf, round2, buffer)

choose_localBestHits(paf, round2)

synteny_paf(
  paf,
  makePlots = 3,
  pdfFile = file.path(getwd(), "dotplots.pdf"),
  ppi = 96,
  minPlotDim = 12,
  maxBestMaps = 5,
  topNhits2keep = 2,
  stepSize,
  windowSize,
  tmpDir,
  MCScanX_hCall,
  blkSize,
  buffer = 5,
  keepEverything = TRUE,
  targetNhitsQuantile = 0.95,
  mapq20nmatchQuantile = 0.1,
  mapq60nmatchQuantile = 0,
  maxDistBtwHitsInBlk = (blkSize * stepSize) * 2,
  verbose = TRUE
)

mcscanx_paf(paf, nGaps, blkSize, MCScanX_hCall, tmpDir, verbose = FALSE)

calc_pafBlkCoords(paf)
}
\arguments{
\item{paf}{character string coercible to a file path specifying the
location of the query (to-be-windowed) fasta file}

\item{mapq20nmatchQuantile}{whats the quantile of nmatches for map20 hits}

\item{mapq60nmatchQuantile}{whats the quantile of nmatches for map60 hits}

\item{round2}{what value should positions be rounded to?}

\item{targetNhitsQuantile}{quantile to choose repetitive regions in the
target}

\item{maxBestMaps}{integer, the maximum number of best-mapping hits to keep}

\item{topNhits2keep}{integer, the number of top hits to keep}

\item{anchorPaf}{paf containing the anchor hits}

\item{buffer}{numeric, position-rank order buffer around anchor hits to
search for additional anchors.}

\item{makePlots}{integer, specifying how many dotplots should be made. O
means no plots. 3 means make one for each step of synteny curation}

\item{ppi}{integer, pixels per inch in the dotplot}

\item{minPlotDim}{numeric, narrowest dimension of the dotplot in inches}

\item{windowSize}{integer, sliding window width}

\item{tmpDir}{character string coercible to a file.path where temporary
files should be written. Cleaned out after run finishes unless this
directory already existed}

\item{MCScanX_hCall}{character string coercible to a file path pointing to
the installed executable for MCScanX_h}

\item{blkSize}{the minimum number of windowed hits to be called a block.}

\item{keepEverything}{logical, should final synteny searches be conducted on
all hits (including masked repeats etc)?}

\item{maxDistBtwHitsInBlk}{numeric, the maximum distance between two
adjacent hits in the same block}

\item{verbose}{logical, should updates be printed to the console?
\cr
If called, \code{synteny_paf} returns its own arguments.}

\item{nGaps}{character string coercible to a file path specifying the
location of the target fasta file}

\item{cleanTmpDir}{logical, should the tmpDir be cleaned out after running?
Always set to TRUE unless manually troubleshooting}

\item{fafiles}{named vector of file.paths to assembly fasta files}

\item{outdir}{file.path to the output directory}

\item{minChrSize}{numeric, minimum chromosome size to be included}

\item{stripChrname}{regex passed to gsub to strip text of chromosome names}

\item{minimap2call}{character string coercible to a valid file path pointing
to the minimap2 program. If installed with conda or in the path, use
"minimap2".}
}
\description{
\code{synteny_paf} methods to constrain windowed hits stored in a paf to
syntenic regions

\code{choose_topNmatch} choose_topNmatch

\code{drop_repetitiveTargets} drop_repetitiveTargets

\code{choose_localUniqHits} choose_localUniqHits

\code{choose_topWindowHits} choose_topWindowHits

\code{project_anchorsOnPaf} project_anchorsOnPaf

\code{choose_localBestHits} choose_localBestHits

\code{synteny_paf} synteny_paf

\code{mcscanx_paf} mcscanx_paf

\code{calc_pafBlkCoords} calc_pafBlkCoords
}
