% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.onAttach}
\alias{prep_fafiles}
\alias{fread_paf}
\alias{convert_pos2ord}
\alias{round_toInteger}
\alias{parse_windPaf}
\alias{calc_linStart}
\alias{parse_pafWindows}
\alias{frank_multicolumn}
\alias{pal_deepspace}
\alias{check_isDNAFasta}
\alias{window_ss}
\alias{set_params}
\alias{interp_xy}
\alias{pull_proxHits}
\alias{check_MCScanXhInstall}
\alias{check_minimap2Install}
\title{Generic internal functions used by DEEPSPACE}
\usage{
.onAttach(...)

prep_fafiles(
  fafiles,
  outdir,
  minChrSize = 0,
  stripChrname = "\\\\s.*",
  nWindows = 1e+06,
  windowSize = 1000,
  stepSize = 1000,
  verbose = TRUE
)

fread_paf(x, verbose = FALSE)

convert_pos2ord(start, end, group, to = 0)

round_toInteger(x, to)

parse_windPaf(paf, queryFastaFile, stripChrname)

calc_linStart(fastaFile = NULL, seqLens = NULL, gapSize = 0)

parse_pafWindows(
  paf,
  querySeqlens,
  addMaprank = TRUE,
  addOrd = TRUE,
  flagRepeat = TRUE,
  addnp = TRUE,
  pidThresh = 0.9,
  ntopThresh = 4,
  npidThresh = ntopThresh * 2,
  round2ord = max(paf$qend - paf$qstart)
)

frank_multicolumn(dt, cols, order, rankWithin = NULL)

pal_deepspace(n)

check_isDNAFasta(faFile)

window_ss(ss, windowSize, windowStep, returnGranges = FALSE)

set_params(
  faFiles,
  wd,
  minChrSize = 0,
  stripChrname = "\\\\s.*",
  nWindows = 1e+06,
  windowSize = 1000,
  stepSize = 1000,
  genomeIDs = NULL,
  refGenome = NULL,
  minimap2.call = "minimap2",
  MCScanX_h.call = "MCScanX_h",
  minimap2.args =
    "-B4 -O6,26 -U50,500 -g100 --no-long-join -r250,500 --rmq=no -A1 -E2,1 -s200 --frag=yes -F1500 -f0 -p0.5 -k19 -w19 -N10",
  nCores = 1,
  blkSize = 5,
  verbose = T
)

interp_xy(x, y)

pull_proxHits(x, y, isAnchor, radius)

check_MCScanXhInstall(filepath)

check_minimap2Install(filepath)
}
\arguments{
\item{fafiles}{named vector of file.paths to assembly fasta files}

\item{outdir}{file.path to the output directory}

\item{minChrSize}{numeric, minimum chromosome size to be included}

\item{stripChrname}{regex passed to gsub to strip text of chromosome names}

\item{windowSize}{integer, sliding window width}

\item{stepSize}{integer, step between sliding window start positions}

\item{verbose}{logical, should updates be printed to the console?
\cr
If called, \code{utils} returns its own arguments.}
}
\description{
\code{utils} Convenience functions for DEEPSPACE, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{.onAttach} startup messages

\code{prep_fafiles} prep_fafiles

\code{fread_paf} Use data.table::fread import the 12 columns of a .paf file

\code{convert_pos2ord} convert_pos2ord

\code{round_toInteger} round_toInteger

\code{parse_windPaf} parse_windPaf

\code{calc_linStart} calc_linStart

\code{parse_pafWindows} parse_pafWindows

\code{frank_multicolumn} frank_multicolumn

\code{pal_deepspace} pal_deepspace

\code{check_isDNAFasta} check_isDNAFasta

\code{window_ss} window_ss

\code{set_params} set_params

\code{interp_xy} interp_xy

\code{pull_proxHits} pull_proxHits

\code{check_MCScanXhInstall} check_MCScanXhInstall

\code{check_minimap2Install} check_minimap2Install
}
